+++ src/main/resources/application.properties
@@ -10,8 +10,8 @@ logging.level.org.springframework.orm.jpa.JpaTransactionManager=DEBUG
 # TODO Append connection to DB
 
 spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://192.168.0.108:7779/shareit
-#spring.datasource.url=jdbc:postgresql://localhost:5432/shareit
+#spring.datasource.url=jdbc:postgresql://192.168.0.108:7779/shareit
+spring.datasource.url=jdbc:postgresql://localhost:5432/shareit
 spring.datasource.username=dbuser
 spring.datasource.password=12345
