+++ src/main/resources/schema.sql
@@ -1,8 +1,4 @@
-drop table if exists users cascade;
-drop table if exists requests cascade;
-drop table if exists items cascade;
-drop table if exists bookings cascade;
-drop table if exists comments cascade;
+
 
 CREATE TABLE IF NOT EXISTS users (
   id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
